+++ h2/connection.py
@@ -38,6 +38,7 @@ from .settings import (
 )
 from .stream import H2Stream
 from .utilities import guard_increment_window
+from .windows import WindowManager
 
 try:
     from hpack.exceptions import OversizedHeaderListError
@@ -360,9 +361,6 @@ class H2Connection(object):
         self.outbound_flow_control_window = (
             self.remote_settings.initial_window_size
         )
-        self.inbound_flow_control_window = (
-            self.local_settings.initial_window_size
-        )
 
         #: The maximum size of a frame that can be emitted by this peer, in
         #: bytes.
@@ -386,6 +384,11 @@ class H2Connection(object):
         # Used to ensure that we don't blow up in the face of frames that were
         # in flight when a stream was reset.
         self._reset_streams = set()
+
+        # The flow control window manager for the connection.
+        self._inbound_flow_control_window_manager = WindowManager(
+            max_window_size=self.local_settings.initial_window_size
+        )
 
         # When in doubt use dict-dispatch.
         self._frame_dispatch_table = {
@@ -481,6 +484,16 @@ class H2Connection(object):
         """
         return self.config.client_side
 
+    @property
+    def inbound_flow_control_window(self):
+        """
+        The size of the inbound flow control window for the connection. This is
+        rarely publicly useful: instead, use :meth:`remote_flow_control_window
+        <h2.connection.H2Connection.remote_flow_control_window>`. This
+        shortcut is largely present to provide a shortcut to this data.
+        """
+        return self._inbound_flow_control_window_manager.current_window_size
+
     def _begin_new_stream(self, stream_id, allowed_ids):
         """
         Initiate a new stream.
@@ -505,13 +518,14 @@ class H2Connection(object):
                 "Invalid stream ID for peer."
             )
 
-        s = H2Stream(stream_id, config=self.config)
+        s = H2Stream(
+            stream_id,
+            config=self.config,
+            inbound_window_size=self.local_settings.initial_window_size,
+            outbound_window_size=self.remote_settings.initial_window_size
+        )
         s.max_inbound_frame_size = self.max_inbound_frame_size
         s.max_outbound_frame_size = self.max_outbound_frame_size
-        s.outbound_flow_control_window = (
-            self.remote_settings.initial_window_size
-        )
-        s.inbound_flow_control_window = self.local_settings.initial_window_size
 
         self.streams[stream_id] = s
 
@@ -879,7 +893,7 @@ class H2Connection(object):
            Rejects attempts to increment the flow control window by out of
            range values with a ``ValueError``.
 
-        :param increment: The amount ot increment the flow control window by.
+        :param increment: The amount to increment the flow control window by.
         :type increment: ``int``
         :param stream_id: (optional) The ID of the stream that should have its
             flow control window opened. If not present or ``None``, the
@@ -901,17 +915,10 @@ class H2Connection(object):
             frames = stream.increase_flow_control_window(
                 increment
             )
-            stream.inbound_flow_control_window = guard_increment_window(
-                stream.inbound_flow_control_window,
-                increment
-            )
         else:
+            self._inbound_flow_control_window_manager.window_opened(increment)
             f = WindowUpdateFrame(0)
             f.window_increment = increment
-            self.inbound_flow_control_window = guard_increment_window(
-                self.inbound_flow_control_window,
-                increment
-            )
             frames = [f]
 
         self._prepare_for_sending(frames)
@@ -1259,6 +1266,56 @@ class H2Connection(object):
             stream.inbound_flow_control_window
         )
 
+    def acknowledge_received_data(self, acknowledged_size, stream_id):
+        """
+        Inform the :class:`H2Connection <h2.connection.H2Connection>` that a
+        certain number of flow-controlled bytes have been processed, and that
+        the space should be handed back to the remote peer at an opportune
+        time.
+
+        .. versionadded:: 2.5.0
+
+        :param acknowledged_size: The total *flow-controlled size* of the data
+            that has been processed. Note that this must include the amount of
+            padding that was sent with that data.
+        :type acknowledged_size: ``int``
+        :param stream_id: The ID of the stream on which this data was received.
+        :type stream_id: ``int``
+        :returns: Nothing
+        :rtype: ``None``
+        """
+        if stream_id <= 0:
+            raise ValueError(
+                "Stream ID %d is not valid for acknowledge_received_data" %
+                stream_id
+            )
+        if acknowledged_size < 0:
+            raise ValueError("Cannot acknowledge negative data")
+
+        frames = []
+
+        conn_manager = self._inbound_flow_control_window_manager
+        conn_increment = conn_manager.process_bytes(acknowledged_size)
+        if conn_increment:
+            f = WindowUpdateFrame(0)
+            f.window_increment = conn_increment
+            frames.append(f)
+
+        try:
+            stream = self._get_stream_by_id(stream_id)
+        except StreamClosedError:
+            # The stream is already gone. We're not worried about incrementing
+            # the window in this case.
+            pass
+        else:
+            # No point incrementing the windows of closed streams.
+            if stream.open:
+                frames.extend(
+                    stream.acknowledge_received_data(acknowledged_size)
+                )
+
+        self._prepare_for_sending(frames)
+
     def data_to_send(self, amt=None):
         """
         Returns some data for sending out of the internal data buffer.
@@ -1361,7 +1418,7 @@ class H2Connection(object):
         delta = new_value - old_value
 
         for stream in self.streams.values():
-            stream.inbound_flow_control_window += delta
+            stream._inbound_flow_control_change_from_settings(delta)
 
     def receive_data(self, data):
         """
@@ -1543,28 +1600,12 @@ class H2Connection(object):
         """
         flow_controlled_length = frame.flow_controlled_length
 
-        try:
-            window_size = self.remote_flow_control_window(frame.stream_id)
-        except NoSuchStreamError:
-            # If the stream doesn't exist we still want to adjust the
-            # connection-level flow control window to keep parity with the
-            # remote peer. If it does exist we'll adjust it later.
-            self.inbound_flow_control_window -= flow_controlled_length
-            raise
-
-        if flow_controlled_length > window_size:
-            raise FlowControlError(
-                "Cannot receive %d bytes, flow control window is %d." %
-                (
-                    flow_controlled_length,
-                    window_size
-                )
-            )
-
         events = self.state_machine.process_input(
             ConnectionInputs.RECV_DATA
         )
-        self.inbound_flow_control_window -= flow_controlled_length
+        self._inbound_flow_control_window_manager.window_consumed(
+            flow_controlled_length
+        )
         stream = self._get_stream_by_id(frame.stream_id)
         frames, stream_events = stream.receive_data(
             frame.data,